Linux hotfix

+++ src/character.rs
@@ -5,5 +5,5 @@ pub struct Character {
     pub texture: Texture,
     pub size: Vector2<i32>,
     pub bearing: Vector2<i32>,
-    pub advance: i32,
+    pub advance: i64,
 }
